refactor: . move fromTask() to EditableTask.ts
